better logic delegation, fixed issues with urls and imports
