+++ daas/daas_app/views/api/get_sample_from_hash.py
@@ -4,8 +4,8 @@ from rest_framework import status
 from drf_yasg.utils import swagger_auto_schema
 from drf_yasg import openapi
 
-from ..models import Sample
-from ..serializers import SampleSerializer
+from ...models import Sample
+from ...serializers import SampleSerializer
 
 
 class GetSampleFromHashAPIView(APIView):
