+++ daas/daas_app/views/api/internals/create_sample.py
@@ -1,22 +1,21 @@
 from rest_framework.response import Response
-from rest_framework.views import APIView
 from rest_framework import status
-from rest_framework.parsers import MultiPartParser, JSONParser
 from drf_yasg.utils import swagger_auto_schema
 from drf_yasg import openapi
-import requests
 import logging
 from typing import List, Optional
 from rest_framework.request import Request
+from rest_framework.views import APIView
 
-from ...utils.callback_manager import CallbackManager
-from ...utils.task_manager import TaskManager
-from ...serializers import SampleSerializer
-from ...utils.lists import recursive_flatten
-from ...models import Sample
+from ...utils.mixins import SampleSubmitMixin
+#from ...utils.callback_manager import CallbackManager
+#from ...utils.task_manager import TaskManager
+from ....serializers import SampleSerializer
+from ....utils.lists import recursive_flatten
+from ....models import Sample
 
 
-class CreateSampleView(APIView):
+class CreateSampleView(SampleSubmitMixin, APIView):
     @swagger_auto_schema(
         operation_id='create_sample',
         request_body=openapi.Schema(
@@ -44,20 +43,20 @@ class CreateSampleView(APIView):
         force_reprocess = request.data.get('force_reprocess', False)
         callback = request.data.get('callback')
         sample_data = request.data.get('sample')
-        samples = self._serialize_sample(sample_data)
+        samples = self._get_and_create_samples(sample_data)
 
-        self._submit_samples(samples)
+        self._submit_samples(samples, force_reprocess)
 
 
         if callback:
             pass  # todo do callback magic here
         return Response(status=status.HTTP_200_OK, data={'non_zip_samples': len(samples)})
 
-    def _serialize_sample(self, sample_data: dict) -> List[Sample]:
+    def _get_and_create_samples(self, sample_data: dict) -> List[Sample]:
         if not sample_data:  # No sample to serialize (either sample not found by meta_extractor or subfiles of an empty zip)
             samples = []
         elif sample_data['file_type'] == 'zip':  # Zip sample
-            samples = [self._serialize_sample(subfile) for subfile in sample_data['subfiles']]
+            samples = [self._get_and_create_samples(subfile) for subfile in sample_data['subfiles']]
         else:  # Non-zip sample
             if sample := self._get_sample(sample_data['sha1']):  # sample already exists
                 samples = [sample]
@@ -70,14 +69,3 @@ class CreateSampleView(APIView):
                     samples = []
                     logging.info(f'Create sample: serializer is not valid: {sample_serializer.data=}')
         return list(set(recursive_flatten(samples)))
-
-    def _submit_samples(self, samples: List[Sample]) -> None:
-        for sample in samples:
-            TaskManager().submit_sample(sample)
-
-    def _get_sample(self, sha1: str) -> Optional[List[Sample]]:
-        try:
-            sample = Sample.objects.get(sha1=sha1)
-        except Sample.DoesNotExist:
-            sample = None
-        return sample
