+++ daas/daas_app/views/reprocess.py
@@ -1,16 +1,13 @@
-from django.urls import reverse
 from django.http import HttpResponseRedirect
-from django.contrib.auth.decorators import login_required
-from django.contrib.auth.decorators import permission_required
-import logging
+from django.shortcuts import reverse
+from django.views import generic
 
+from .utils.mixins import ReprocessMixin
 from ..models import Sample
-from ..utils.reprocess import reprocess
 
 
-@login_required
-@permission_required('upload_sample_permission')
-def reprocess_view(request, sample_id):
-    logging.info('Reprocessing sample: %s' % sample_id)
-    reprocess(Sample.objects.get(id=sample_id), force_reprocess=True)
-    return HttpResponseRedirect(reverse('index'))
+class ReprocessWebView(ReprocessMixin, generic.View):
+    def get(self, request, sample_id: int):
+        self.reprocess(Sample.objects.filter(id=sample_id))
+
+        return HttpResponseRedirect(reverse('index'))
